requirements.txt
appdirs==1.4.3
asgiref==3.2.7
backcall==0.2.0
certifi==2020.4.5.1
cfgv==3.2.0
chardet==3.0.4
decorator==4.4.2
distlib==0.3.1
Django==3.0.4
django-autocomplete-light==3.5.1
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-meta==1.6.1
django-recaptcha3==0.2.0
django-tag-parser==3.2
django-taggit==1.2.0
dnspython==1.16.0
docopt==0.6.2
feedparser==5.2.1
filelock==3.0.12
flake8==3.8.3
identify==1.4.30
idna==2.9
importlib-metadata==1.7.0
ipdb==0.13.3
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
mccabe==0.6.1
nodeenv==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py3-validate-email==0.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
uritools==3.0.0
urlextract==0.14.0
urllib3==1.25.9
virtualenv==20.0.31
wcwidth==0.2.5
zipp==3.1.0
